python/requirements.txt
faster-whisper>=1.0.3
openwakeword>=0.6.0
kokoro-onnx
fastmcp
numpy>=1.26.0
webrtcvad>=2.0.10
pyaudio>=0.2.14
soundfile
